static commands: remove magic expression reordering
preparation for awaits in client-side expressions
