+++ src/DotVVM.Framework/Compilation/Javascript/ResultIsPromiseAnnotation.cs
@@ -10,6 +10,7 @@ namespace DotVVM.Framework.Compilation.Javascript
         public Func<JsExpression, JsExpression> GetPromiseFromExpression;
         public List<object> ResultAnnotations;
 
+        public ResultIsPromiseAnnotation(params object[] resultAnnotations): this(e => e, resultAnnotations) { }
         public ResultIsPromiseAnnotation(Func<JsExpression, JsExpression> promiseGetter, params object[] resultAnnotations)
         {
             this.ResultAnnotations = resultAnnotations.ToList();
