+++ src/DotVVM.Framework/DependencyInjection/DotVVMServiceCollectionExtensions.cs
@@ -79,6 +79,7 @@ namespace Microsoft.Extensions.DependencyInjection
             services.TryAddScoped<RuntimeWarningCollector>();
             services.TryAddScoped<AggregateRequestTracer, AggregateRequestTracer>();
             services.TryAddScoped<ResourceManager, ResourceManager>();
+            services.TryAddSingleton<StaticCommandMethodTranslator>();
             services.TryAddSingleton(s => DotvvmConfiguration.CreateDefault(s));
             services.TryAddSingleton(s => s.GetRequiredService<DotvvmConfiguration>().Markup);
             services.TryAddSingleton(s => s.GetRequiredService<DotvvmConfiguration>().Resources);
